feat: itens negociados em toda view de nota + parecer verde no clearance
Itens da nota em todas as superfícies que exibem uma NF-e/EFD:
- Partial canônico `_itens-tabela` (XML/EFD, null-safe, teto/truncamento)
- Expansivo `/app/clearance/notas`, detalhe `/app/clearance/nota/{id}`,
  `xml-inline` (cobre /app/notas + cards de cliente/participante/empresa),
  `xml-nota` e resultado do lote `/app/clearance/notas/resultado/{id}`
  (itens por chave, acervo XML preferido + fallback gêmea EFD Contribuições)
- Parecer positivo do clearance em verde (era cinza)
- Histórico de monitoramento: badge de certidão canônico (CertidaoBadge) no
  backend, front vira render burro com as 6 fontes

Migration `monitoramento_planos.custo_creditos` integer→decimal(8,2) para
comportar o preço em R$ do re-base.

O tree também carrega o re-base de saldo para R$ (workstream paralelo, já
concluído e rodado em prod) e contas multiusuário. Suíte Clearance +
Monitoramento + Unit verde (554 passed).

+++ app/Jobs/ProcessarAutoTopUpJob.php
@@ -47,7 +47,7 @@ class ProcessarAutoTopUpJob implements ShouldQueue
             }
 
             $user = User::find($this->userId);
-            if ($user === null || (int) $user->credits >= (int) $r->limite_creditos) {
+            if ($user === null || (float) $user->credits >= (float) $r->limite_creditos) {
                 return null;
             }
